moved types to model and changed Item type
