+++ apps/frontend/src/app/common/table/table.component.stories.ts
@@ -33,7 +33,7 @@ const template: Story<TableComponent> = args => ({
     ...args,
     buttonClicked: event => {
       console.log(
-        `handle button click event: ${event.buttonType} / ${event.itemId}`
+        `handle button click event: ${event.buttonType} / ${event.item.id}`
       );
     }
   },
